% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-helpers-text.R
\name{clean_text}
\alias{clean_text}
\title{Clean up text}
\usage{
clean_text(string, expand_contractions = TRUE)
}
\arguments{
\item{string}{Length-1 character}

\item{expand_contractions}{Whether to expand contractions. Defaults to \code{TRUE}}
}
\value{
Character
}
\description{
Clean up text
}
