% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-analyze-metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Extract metrics from match matrix}
\usage{
get_metrics(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
List
}
\description{
Extract metrics from match matrix
}
