% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-match_matrix.R
\name{match_matrix}
\alias{match_matrix}
\title{Calculate match matrix}
\usage{
match_matrix(hypothesis, reference, unit = c("letter", "phon", "none"))
}
\arguments{
\item{hypothesis}{Observed string of text}

\item{reference}{True string of text}

\item{unit}{How edit distance should be calculated, if at all}
}
\value{
Matrix
}
\description{
Calculate match matrix
}
