% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-data.R
\docType{data}
\name{stella_observed}
\alias{stella_observed}
\title{Transcribed text of the "Please call Stella" passage}
\format{
A character vector
}
\source{
\url{https://accent.gmu.edu/howto.php}
}
\usage{
stella_observed
}
\description{
Transcribed text of the "Please call Stella" passage
}
\keyword{datasets}
