% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-data.R
\docType{data}
\name{stella_truth}
\alias{stella_truth}
\title{True text of the "Please call Stella" passage}
\format{
A character vector
}
\source{
\url{https://accent.gmu.edu/howto.php}
}
\usage{
stella_truth
}
\description{
True text of the "Please call Stella" passage
}
\keyword{datasets}
