% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-match_matrix.R
\name{match_matrix}
\alias{match_matrix}
\title{Calculate match matrix}
\usage{
match_matrix(
  hypothesis,
  reference,
  clean = TRUE,
  unit = c("word", "letter", "phon", "none"),
  out = c("matrix", "edits")
)
}
\arguments{
\item{hypothesis}{Observed string of text}

\item{reference}{True string of text}

\item{clean}{Whether to apply \code{clean_text()} to the text inputs}

\item{unit}{How edit distance should be calculated, if at all}

\item{out}{Whether to return the matrix (\code{"matrix"}) or a dataframe of edits (\code{"edit"})}
}
\value{
Matrix
}
\description{
Calculate match matrix
}
